Use AudioTrack instead of MediaPlayer for playback
MediaPlayer was not able to loop the audio without experiencing
a gap. A more direct approach is necessary to clear up the gap.

This commit uses FFmpeg to convert the mp3 file into a PCM file.
From there, it is read directly and applied to an AudioTrack
in a loop.