Set the default value of the low pass filter to 1000 Hz and max to 5000 Hz
The limit of human hearing is ~20kHz. This is the reason for the 20000 Hz
limit on the low pass filter. However, the ear is most sensitive around
1kHz-4kHz.

A default value of 10kHz for the filter may be a bit too high to do
anything meaningful. For some examples, FFmpeg sets a default value
of 500Hz and Audacity's default is 1000Hz. This commit sets the
default value to 1000Hz.

Further, the maximum value of the filter is too high to be
useful. For some of the samples there is no much change above
10kHz. Setting the max value around 5000Hz.