Simplify playback to one thread that loops over one AudioTrack
The previous implementation had a flaw, in that it assumed the
two threads would always stay at the correct phase of each other.
This would be enforced by the callback from the AudioTrack
informing of a specific place in playback was reached. The issue
with this is the callback would not always occur at the same time,
and over the course of several hours the two threads would
near synchronize with each other, leading to a playback
that faded in and out.

The reason for that implementation was the concern that it would
not be possible to play seamlessly with one AudioTrack. In early
prototypes there was a rough 'popping' noise when playback started.
However, it turned out that this was because the initial file
played was a PCM file, which includes header data. That header data
was mistakenly being played as sound, leading to the popping.
As a RAW format is now used, there is no popping.

With this in mind, this commit removes the multiple threads approach
to playback and now uses only one thread which will play a file
to completion they start again with the same AudioTrack.

The audio samples themselves were modified to seamlessly loop into
themselves. Namely, the following transformations were done using
Audacity:
 - take a ~10 second clip
 - Apply a 2 second fade-in and fade-out
 - Cut the first 3 seconds, and place it over the last three seconds
   which should create a seamless track appropriate for looping
 - Save as a mp3 file, 128kbps, stereo

Finally, in addition to this, a wake lock is now being held during
playback to prevent the phone from shutting off the CPU during playback.

https://github.com/brarcher/baby-sleep-sounds/issues/30