Play audio from a foreground service
Android will periodically look through the list of apps which
are running in the background. If they are holding a wakelock
the app is killed.

This app would hold a wakelock, but was doing so from an
Activity. When the Activity goes to the background Android
believes the wakelock hold is in error and does not honor it.
To get around this, the audio playback is moved to a Service.
The Service registers as a foreground service. Then, when
it holds a lock Android will not attempt to kill it.